Refactor frontend completely
- Rename components to make more sense
- Move global state to App and move code from monolith file to new
components for modularity
- Consolidate and simplify some components into a single component
- Update layout, styling, transitions, and animations

Initial commit with stats included and context expected

Update references to TypingArea

Import Results component and add into rendered output

Incorporate context into app

Move all state to App

Messy refactor
