+++ frontend/src/components/Button.jsx
@@ -1,12 +1,23 @@
-export default function WordListButton({ text, list, onClick, isActive }) {
+export default function Button({
+  onClick,
+  isActive,
+  text = null,
+  mode = null,
+  group = null,
+  duration = null,
+  count = null,
+}) {
   return (
     <>
       {isActive ? (
         <button
           type="button"
           className="bg-sky-950 text-sky-200 p-2"
-          list={list}
           onClick={onClick}
+          mode={mode}
+          group={group}
+          duration={duration}
+          count={count}
         >
           {text}
         </button>
@@ -14,8 +25,11 @@ export default function WordListButton({ text, list, onClick, isActive }) {
         <button
           type="button"
           className="bg-sky-950 text-sky-300/50 hover:text-sky-200 p-2"
-          list={list}
           onClick={onClick}
+          mode={mode}
+          group={group}
+          duration={duration}
+          count={count}
         >
           {text}
         </button>
