+++ frontend/src/components/Header.jsx
@@ -3,7 +3,7 @@ export default function Header() {
     <>
       <header
         className="flex w-full
-      max-w-3xl justify-between items-center px-4 py-4"
+      max-w-3xl justify-between items-center m-4"
       >
         <div className="flex gap-4">
           <h1 className="text-4xl">Yubi</h1>
